Extracts useChat composable
